requirements.txt
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-default-group==1.2.4
distro==1.9.0
duckduckgo_search==8.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
openai==2.14.0
playwright==1.57.0
playwright-stealth==2.0.0
pluggy==1.6.0
primp==0.15.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlite-fts4==1.0.3
sqlite-utils==3.39
tabulate==0.9.0
tavily-python==0.7.17
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
websockets==15.0.1
